em processo

+++ curso_java/Atividades gpt/loja_produtos/Produto.java
@@ -3,7 +3,7 @@ package loja_produtos;
 
 public class Produto{
     public String nome;
-    public int preco;
+    public double preco;
 
     public Produto(String nome, double preco){
         this.nome = nome;
